getNextActiveCellUseCase code update
