+++ src/core/EULA.ts
@@ -1,0 +1,10 @@
+export const EULAContent = `
+User Agreement
+Algorithm content note update: The algorithm has been developed to calculate any test including those with one or more Not Testable (NT) values. This algorithm should not be used for individuals who have a SCI at more than one level (i.e., two cord injuries at different levels).
+
+The algorithm will work with most browsers, however some esthetic features will not appear with earlier versions of Internet Explorer.
+
+Legal: The information provided on this web site is for educational and informational purposes only. You use this website and the information it provides on an "as is" basis. No representations or warranties are made as to its accuracy. This information should not be regarded as complete or a source of medical or diagnostic information or used as a substitute for clinical judgment of professional medical instruction. This program was developed by the Praxis Spinal Cord Institute (Praxis) based on the American Spinal Injury Association's (ASIA) International Standards for Neurological Classification of Spinal Cord Injury Revised 2019. All content accessible on this site is being provided without liability on the part of Praxis and no representation or warranty with respect to the content is being made by ASIA or the International Spinal Cord Society (ISCoS). Praxis, ASIA, and ISCoS disclaim all representations, warranties, conditions, and guarantees, whether expressed or implied, statutory or otherwise, with respect to this website and the information it provides, including any representations, warranties, conditions, and guarantees regarding non-infringement, merchantability, fitness for a particular purpose, or that arise under any sales of goods legislation such as the United Nations Convention on contracts for the International Sale of Goods.
+
+In no event shall Praxis, ASIA, or ISCoS be liable to you or any other entity for any direct, special, consequential, incidental, or indirect damages, liabilities, losses (including lost profits), costs (including legal fees and disbursements), expenses, claims, fines, penalties, demands, suits, actions, proceedings, or judgments, however caused, on any theory of liability, and notwithstanding any failure of essential purpose.
+`;
