+++ src/core/useCases/index.ts
@@ -8,3 +8,4 @@ export {setExtraInputsUseCase} from './setExtraInputs.useCase';
 export {setReadonlyUseCase} from './setReadonly.useCase';
 export {setStarDetailsUseCase} from './setStarDetails.useCase';
 export {setVacDapUseCase} from './setVacDap.useCase';
+export {getNextActiveCellUseCase} from './getNextActiveCell.useCase';
